moved checkbox component to shared folder
