+++ src/components/ProductListPage/CheckBoxList/CheckBoxList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import CheckBox from './CheckBox';
+import CheckBox from '../../Shared/CheckBox';
 
 const CheckBoxList = ({ itemsList, category, checkCheckbox }) => {
   const renderCheckboxes = itemsList.map(({ item }) => (
